+++ src/py42/services/orgs.py
@@ -1,10 +1,9 @@
 from collections import namedtuple
 
-from py42 import settings
 from py42.clients.settings.org_settings import OrgSettings
 from py42.exceptions import Py42Error
+from py42.exceptions import Py42InternalServerError
 from py42.services import BaseService
-from py42.services.util import get_all_pages
 
 OrgSettingsResponse = namedtuple(
     "OrgSettingsResponse", ["error", "org_response", "org_settings_response"]
@@ -17,6 +16,20 @@ class OrgService(BaseService):
     Use the OrgService to create and retrieve organizations. You can also use it to block and
     deactivate organizations.
     """
+
+    def __init__(self, connection):
+        super().__init__(connection)
+        self._org_id_map = {}
+
+    @property
+    def org_id_map(self):
+        """Map org guids to ids."""
+        if not self._org_id_map:
+            self._org_id_map = {}
+            page = self.get_page()
+            for org in page["orgs"]:
+                self._org_id_map[org["orgId"]] = org["orgGuid"]
+        return self._org_id_map
 
     def create_org(self, org_name, org_ext_ref=None, notes=None, parent_org_uid=None):
         """Creates a new organization.
@@ -33,14 +46,23 @@ class OrgService(BaseService):
         Returns:
             :class:`py42.response.Py42Response`
         """
-        uri = "/api/v1/Org/"
+
+        # get parent org GUID from UID :)
+        parent_org_guid = self._get_guid_by_id(parent_org_uid, id_key="orgUid")
+
+        uri = "/api/v3/orgs"
         data = {
             "orgName": org_name,
             "orgExtRef": org_ext_ref,
             "notes": notes,
-            "parentOrgUid": parent_org_uid,
+            "parentOrgGuid": parent_org_guid,
         }
-        return self._connection.post(uri, json=data)
+        response = self._connection.post(uri, json=data)
+
+        # update ID store
+        self.org_id_map[response["orgId"]] = response["orgGuid"]
+
+        return response
 
     def get_by_id(self, org_id, **kwargs):
         """Gets the organization with the given ID.
@@ -51,7 +73,7 @@ class OrgService(BaseService):
         Returns:
             :class:`py42.response.Py42Response`: A response containing the organization.
         """
-        uri = f"/api/v1/Org/{org_id}"
+        uri = f"/api/v3/orgs/{self._get_guid_by_id(org_id)}"
         return self._connection.get(uri, params=kwargs)
 
     def get_by_uid(self, org_uid, **kwargs):
@@ -63,26 +85,21 @@ class OrgService(BaseService):
         Returns:
             :class:`py42.response.Py42Response`: A response containing the organization.
         """
-        uri = f"/api/v1/Org/{org_uid}"
-        params = dict(idType="orgUid", **kwargs)
-        return self._connection.get(uri, params=params)
+        uri = f'/api/v3/orgs/{self._get_guid_by_id(org_uid, id_key="orgUid")}'
+        return self._connection.get(uri, params=kwargs)
 
-    def get_page(self, page_num, page_size=None, **kwargs):
+    def get_page(self, page_num=1, page_size=None, **kwargs):
         """Gets an individual page of organizations.
 
         Args:
-            page_num (int): The page number to request.
-            page_size (int, optional): The number of organizations to return per page.
-                Defaults to `py42.settings.items_per_page`.
-            kwargs (dict, optional): Additional advanced-user arguments. Defaults to None.
+            page_num (int, optional): DEPRECATED.
+            page_size (int, optional): DEPRECATED.
 
         Returns:
             :class:`py42.response.Py42Response`
         """
-        page_size = page_size or settings.items_per_page
-        uri = "/api/v1/Org"
-        params = dict(pgNum=page_num, pgSize=page_size, **kwargs)
-        return self._connection.get(uri, params=params)
+        uri = "/api/v3/orgs"
+        return self._connection.get(uri)
 
     def get_all(self, **kwargs):
         """Gets all organizations.
@@ -91,7 +108,7 @@ class OrgService(BaseService):
             generator: An object that iterates over :class:`py42.response.Py42Response` objects
             that each contain a page of organizations.
         """
-        return get_all_pages(self.get_page, "orgs", **kwargs)
+        yield self.get_page()
 
     def block(self, org_id):
         """Blocks the organization with the given org ID as well as its child organizations. A
@@ -105,8 +122,8 @@ class OrgService(BaseService):
         Returns:
             :class:`py42.response.Py42Response`
         """
-        uri = f"/api/v1/OrgBlock/{org_id}"
-        return self._connection.put(uri)
+        uri = f"/api/v3/orgs/{self._get_guid_by_id(org_id)}/block"
+        return self._connection.post(uri)
 
     def unblock(self, org_id):
         """Removes a block, if one exists, on an organization and its descendants with the given
@@ -118,8 +135,8 @@ class OrgService(BaseService):
         Returns:
             :class:`py42.response.Py42Response`
         """
-        uri = f"/api/v1/OrgBlock/{org_id}"
-        return self._connection.delete(uri)
+        uri = f"/api/v3/orgs/{self._get_guid_by_id(org_id)}/unblock"
+        return self._connection.post(uri)
 
     def deactivate(self, org_id):
         """Deactivates the organization with the given ID, including all users, plans, and
@@ -131,8 +148,8 @@ class OrgService(BaseService):
         Returns:
             :class:`py42.response.Py42Response`
         """
-        uri = f"/api/v1/OrgDeactivation/{org_id}"
-        return self._connection.put(uri)
+        uri = f"/api/v3/orgs/{self._get_guid_by_id(org_id)}/deactivate"
+        return self._connection.post(uri)
 
     def reactivate(self, org_id):
         """Reactivates the organization with the given ID. Backups are *not* restarted
@@ -144,18 +161,26 @@ class OrgService(BaseService):
         Returns:
             :class:`py42.response.Py42Response`
         """
-        uri = f"/api/v1/OrgDeactivation/{org_id}"
-        return self._connection.delete(uri)
+        uri = f"/api/v3/orgs/{self._get_guid_by_id(org_id)}/activate"
+        return self._connection.post(uri)
 
     def get_current(self, **kwargs):
         """Gets the organization for the currently signed-in user.
+
+        WARNING: This method is incompatible with api client authentication.
 
         Returns:
             :class:`py42.response.Py42Response`: A response containing the organization for the
             currently signed-in user.
         """
         uri = "/api/v1/Org/my"
-        return self._connection.get(uri, params=kwargs)
+        try:
+            return self._connection.get(uri, params=kwargs)
+        except Py42InternalServerError as err:
+            raise Py42InternalServerError(
+                err,
+                message="Server Error. Please be aware that this method is incompatible with api client authentication.",
+            )
 
     def get_agent_state(self, org_id, property_name):
         """Gets the agent state of the devices in the org.
@@ -232,3 +257,39 @@ class OrgService(BaseService):
             org_response=org_response,
             org_settings_response=org_settings_response,
         )
+
+    def update_org(self, org_id, name=None, notes=None, ext_ref=None):
+        """Updates an org. Only fields associated with passed parameters will be updated.
+
+
+        Args:
+            org_id (str): The org's identifier.
+            name (str, optional): The updated name for the org.
+            notes (str, optional): The updated notes for the org.
+            ext_ref (str, optional): The updated external reference for the org.
+
+        Returns:
+            :class:`py42.response.Py42Response`:
+        """
+        uri = f"/api/v3/orgs/{self._get_guid_by_id(org_id)}"
+        data = {"orgName": name, "orgExtRef": ext_ref, "notes": notes}
+        self._connection.put(uri, json=data)
+
+    def _get_guid_by_id(self, org_id, id_key="orgId"):
+        # Identity crisis helper method.
+        # Old orgs methods accepted IDs. New apis take GUIDs.
+        # Use additional lookup to prevent breaking changes.
+
+        if id_key != "orgId":
+            guid = ""
+            page = self.get_page()
+            for org in page["orgs"]:
+                if org[id_key] == org_id:
+                    return org["orgGuid"]
+            if not guid:
+                raise Py42Error(f"Couldn't find an Org with ID '{org_id}'.")
+        else:
+            try:
+                return self.org_id_map[org_id]
+            except KeyError:
+                raise Py42Error(f"Couldn't find an Org with ID '{org_id}'.")